Improve the performance of softmax in the mid range
We choose to go for a large block size. In any of the casees, the number of blocks per SM is 1,
so this increases the parallelism. Also, the PTX code seems to be more efficient
(the memory throughput goes from 69% to 77%). To go with this, we also increment the max RBLOCK.

ghstack-source-id: f9429438ad4681c58dbf0445ff90815b5581bdac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/90994